+++ slumber-bard/src/main/java/io/samituga/bard/endpoint/type/QueryParamName.java
@@ -1,6 +1,6 @@
 package io.samituga.bard.endpoint.type;
 
-import io.samituga.slumber.heimer.type.Type;
+import io.samituga.slumber.ivern.type.Type;
 
 public class QueryParamName extends Type<String> {
     private QueryParamName(String value) {
