+++ slumber-bard/src/main/java/module-info.java
@@ -9,4 +9,5 @@ module slumber.bard {
 
     requires jakarta.servlet;
     requires slumber.heimer;
+    requires slumber.ivern;
 }
