+++ slumber-bard/src/test/java/io/samituga/bard/fixture/FilterTestDataTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.samituga.bard.endpoint.type.Path;
 import io.samituga.bard.filter.type.Order;
-import io.samituga.slumber.heimer.exception.AssertionException;
+import io.samituga.slumber.heimer.validator.exception.AssertionException;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import org.junit.jupiter.api.Test;
