+++ slumber-bard/src/test/java/io/samituga/bard/fixture/ResponseTestDataTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.samituga.bard.endpoint.HttpCode;
-import io.samituga.slumber.heimer.exception.AssertionException;
+import io.samituga.slumber.heimer.validator.exception.AssertionException;
 import org.junit.jupiter.api.Test;
 
 class ResponseTestDataTest {
