+++ slumber-heimer/src/main/java/module-info.java
@@ -1,5 +1,3 @@
 module slumber.heimer {
-    exports io.samituga.slumber.heimer.builder;
     exports io.samituga.slumber.heimer.validator;
-    exports io.samituga.slumber.heimer.type;
 }
