+++ slumber-ivern/src/main/java/io/samituga/slumber/ivern/builder/Builder.java
@@ -1,4 +1,4 @@
-package io.samituga.slumber.heimer.builder;
+package io.samituga.slumber.ivern.builder;
 
 public interface Builder<T> {
     T build();
