+++ slumber-ivern/src/main/java/io/samituga/slumber/ivern/type/Type.java
@@ -1,4 +1,4 @@
-package io.samituga.slumber.heimer.type;
+package io.samituga.slumber.ivern.type;
 
 import java.util.Objects;
