+++ slumber-ivern/src/main/java/module-info.java
@@ -1,3 +1,6 @@
 module slumber.ivern {
-    exports io.samituga.slumber.ivern.repository;
+    exports io.samituga.slumber.ivern.data.repository;
+    exports io.samituga.slumber.ivern.http.type;
+    exports io.samituga.slumber.ivern.builder;
+    exports io.samituga.slumber.ivern.type;
 }
