+++ slumber-malz-jooq/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module slumber.malz.jooq {
     exports io.samituga.slumber.malz.jooq.factory;
 
     requires slumber.heimer;
+    requires slumber.ivern;
     requires slumber.malz;
     requires org.jooq;
     requires com.zaxxer.hikari;
