+++ slumber-malz-jooq/src/test/java/io/samituga/slumber/malz/jooq/config/DataSourceConfigTest.java
@@ -4,7 +4,7 @@ import static io.samituga.slumber.malz.driver.Driver.POSTGRES;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.samituga.slumber.heimer.exception.AssertionException;
+import io.samituga.slumber.heimer.validator.exception.AssertionException;
 import io.samituga.slumber.malz.driver.Driver;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
